feat: make the capability traits implementable
- Thirteen were pub(crate), so only three of sixteen capabilities
  could be implemented downstream. The four already public carried
  doc(hidden) and a note calling them a bench seam.
- Re-exports async_trait, since every impl needs the attribute and a
  downstream adding its own risks a version mismatch.
- ProviderAdapter drops #[allow(private_interfaces)]: the leak it
  silenced was the capability traits themselves.
- Splits adapter.rs along the seam the traits already fall on, equity
  versus market-wide versus dispatch.
